% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knor.R
\name{Gmeans}
\alias{Gmeans}
\title{Perform a parallel hierarchical clustering using the g-means algorithm}
\usage{
Gmeans(data, kmax, nrow = -1, ncol = -1, iter.max = 20,
  nthread = -1, init = c("forgy"), tolerance = 1e-06,
  dist.type = c("eucl", "cos", "taxi"), min.clust.size = 1,
  strictness = 4)
}
\arguments{
\item{data}{Data file name on disk (NUMA optmized) or In memory data matrix}

\item{kmax}{The maximum number of centers}

\item{nrow}{The number of samples in the dataset}

\item{ncol}{The number of features in the dataset}

\item{iter.max}{The maximum number of iteration of k-means to perform}

\item{nthread}{The number of parallel threads to run}

\item{init}{The type of initialization to use c("forgy") or initial centers}

\item{tolerance}{The convergence tolerance for k-means at each
hierarchical split}

\item{dist.type}{What dissimilarity metric to use}

\item{min.clust.size}{The minimum size of a cluster when it cannot be split}

\item{strictness}{The Anderson-Darling strictness level.
Should be between 1 and 4 inclusive}
}
\value{
A list of lists containing the attributes of the output of kmeans.
 cluster: A vector of integers (from 1:\strong{k}) indicating the cluster to
         which each point is allocated.
 centers: A matrix of cluster centres.
 size: The number of points in each cluster.
 iter: The number of (outer) iterations.
}
\description{
A hierarchical cluster algorithm that chooses the number of clusters based on
 the Anderson Darling statistic described in:
 http://papers.nips.cc/paper/2526-learning-the-k-in-k-means.pdf
}
\examples{
iris.mat <- as.matrix(iris[,1:4])
kmax <- length(unique(iris[, dim(iris)[2]])) # Number of unique classes
xms <- Gmeans(iris.mat, kmax)

}
\author{
Disa Mhembere <disa@cs.jhu.edu>
}
