% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knor.R
\name{GMM}
\alias{GMM}
\title{Fit a mixture of Gaussians distribution to the data.}
\usage{
GMM(data, nrow = -1, ncol = -1, k, reg.covar = 1e-06,
  iter.max = 100, nthread = -1, tolerance = 1e-06,
  cov.type = c("full", "tied", "diag", "spherical"))
}
\arguments{
\item{data}{Data file name on disk (NUMA optimized) or In memory data matrix}

\item{nrow}{The number of samples in the dataset}

\item{ncol}{The number of features in the dataset}

\item{k}{The number of gaussians to estimate}

\item{reg.covar}{The covariance matrix regularization term}

\item{iter.max}{The maximum number of iteration of k-means to perform}

\item{nthread}{The number of parallel thread to run}

\item{tolerance}{The convergence tolerance}

\item{cov.type}{is the type of covariance matrix. It can be
one of {"full", "tied", "diag", "spherical"}.
\itemize{
\item{"full"}{each component has its own general covariance matrix.}
\item{"tied"}{all components share the same general covariance matrix.}
\item{"diag"}{each component has its own diagonal covariance matrix.}
\item{"spherical"}{each component has its own single variance.}
}}
}
\value{
A list containing the attributes of the output of GMM.
      \itemize{
       \item{loglik}{a n x k matrix, whose \code{[i, k]}th entry is
               the conditional probability of the ith observation
               belonging to the kth component of the mixture.}
       \item{iter}{the number of iterations}
       \item{parameters}{parameters of the mixture of Gaussian distribution.
            \itemize{
            \item{weights}{a vector of k elements. Each element is
             the weight of the Gaussian distribution in the mixture.}
            \item{means}{a k x d matrix. Each row is the mean of a
              Gaussian distribution.}
            \item{covs}{a list of matrices, a matrix or a vector,
              depending on \code{cov.type}}}
       }
     }
}
\description{
Fit a mixture of Gaussians distribution to the data.
}
\examples{
iris.mat <- as.matrix(iris[,1:4])
k <- length(unique(iris[, dim(iris)[2]])) # Number of unique classes
kms <- GMM(iris.mat, k, cov=c("full"))

}
\author{
Disa Mhembere <disa@cs.jhu.edu>
}
