% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knor.R
\name{kmeans}
\alias{kmeans}
\title{Perform k-means clustering on a data matrix.}
\usage{
kmeans(datafn, nrow, ncol, k, max.iters = .Machine$integer.max,
  nthread = -1, centers = NULL, init = c("kmeanspp", "random", "forgy",
  "none"), tolerance = 1e-06, dist.type = c("eucl", "cos"),
  centersfn = "", omp = FALSE)
}
\arguments{
\item{datafn}{Data file name on disk}

\item{nrow}{The number of samples in the dataset}

\item{ncol}{The number of features in the dataset}

\item{k}{The desired number of clusters}

\item{max.iters}{Then maximum number of iteration of k-means to perform}

\item{nthread}{The number of parallel thread to run}

\item{centers}{Pre-computed centroids}

\item{init}{The type of initialization to use}

\item{tolerance}{The convergence tolerance}

\item{dist.type}{What dissimilarity metric to use}

\item{centersfn}{Filename containing centers}

\item{omp}{Use (slower) OpenMP threads rather than pthreads (default: FALSE)}
}
\value{
A list containing the output of kmeans.
}
\description{
K-means provides `k` disjoint sets for a dataset.
}
\author{
Disa Mhembere <disa@jhu.edu>
}
