% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_centroids}
\alias{test_centroids}
\title{A small example of centroids of dim: (8,5) used as for micro-benchmarks
of the clusternor package. The data are randomly generated.}
\format{An object of class \code{"matrix"}}
\usage{
data(test_centroids)
}
\description{
A small example of centroids of dim: (8,5) used as for micro-benchmarks
of the clusternor package. The data are randomly generated.
}
\examples{
data(test_centroids)
kms <- Kmeans(test_data, test_centroids)
}
\keyword{datasets}
