% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{A small dataset of dim: (50,5) used as for micro-benchmarks of the knor
package. The data are randomly generated hence a clear number of clusters
will be hard to find.}
\format{An object of class \code{"matrix"}}
\usage{
data(test_data)
}
\description{
A small dataset of dim: (50,5) used as for micro-benchmarks of the knor
package. The data are randomly generated hence a clear number of clusters
will be hard to find.
}
\examples{
ncenters <- 8
kms <- Kmeans(test_data, ncenters)
}
\keyword{datasets}
