% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knor.R
\name{kmeans}
\alias{kmeans}
\title{Perform k-means clustering on a data matrix.}
\usage{
kmeans(data, centers, nrow = -1, ncol = -1,
  max.iters = .Machine$integer.max, nthread = -1, init = c("kmeanspp",
  "random", "forgy", "none"), tolerance = 1e-06, dist.type = c("eucl",
  "cos"), omp = FALSE, numa.opt = FALSE)
}
\arguments{
\item{data}{Data file name on disk or In memory data matrix}

\item{centers}{Either (i) The number of centers (i.e., k), or
(ii) an In-memory data matrix, or (iii) A 2-Element `list` with element 1
being a filename for precomputed centers, and element 2
the number of centroids.}

\item{nrow}{The number of samples in the dataset}

\item{ncol}{The number of features in the dataset}

\item{max.iters}{Then maximum number of iteration of k-means to perform}

\item{nthread}{The number of parallel thread to run}

\item{init}{The type of initialization to use c("kmeanspp", "random",
"forgy", "none")}

\item{tolerance}{The convergence tolerance}

\item{dist.type}{What dissimilarity metric to use c("eucl", "cos")}

\item{omp}{Use (slower) OpenMP threads rather than pthreads (default: FALSE)}

\item{numa.opt}{When passing `data` as an in-memory data matrix you can
optimize memory placement for Linux NUMA machines. **NOTE** that performance may
degrade with very large data & it requires 2*memory of that without this.}
}
\value{
A list containing the attributes of the output of kmeans.
 cluster: A vector of integers (from ‘1:k’) indicating the cluster to
         which each point is allocated.
 centers: A matrix of cluster centres.
 size: The number of points in each cluster.
 iter: The number of (outer) iterations.
}
\description{
K-means provides `k` disjoint sets for a dataset using a parallel and fast
NUMA optimized version of Lloyd's algorithm. The details of which are found
in this paper https://arxiv.org/pdf/1606.08905.pdf.
}
\author{
Disa Mhembere <disa@jhu.edu>
}
